Refactor begins on WidgetHandler
Using alternative to deprecated classes and methods
commented out skill and spell functionality.
skills used to require have a new widget created to act as a cover, however game updates have now allowed to act directly on the existing skill widgets. so a refactor is needed to bring it in line with the prayer widget method